@@ -1,9 +1,9 @@
-package com.globant.counter.android.mvp.view;
+package com.globant.mvpcounter.mvp.view;
 
 import android.app.Activity;
 import android.app.FragmentManager;
 import android.content.Context;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import java.lang.ref.WeakReference;
 
