@@ -1,8 +1,7 @@
-package com.globant.counter.android.mvp.presenter;
+package com.globant.mvpcounter.mvp.presenter;
 
-import com.globant.counter.android.mvp.model.CountModel;
-import com.globant.counter.android.mvp.view.CountView;
-
+import com.globant.mvpcounter.mvp.model.CountModel;
+import com.globant.mvpcounter.mvp.view.CountView;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -12,12 +11,6 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
 public class PresenterTest {
 
     private CountPresenter presenter;
@@ -32,14 +25,14 @@ public class PresenterTest {
     }
 
     @Test
-    public void addition_isCorrect() throws Exception {
+    public void addition_isCorrect() {
         assertEquals(4, 2 + 2);
     }
 
     @Test
     public void isShouldIncCountByOne() {
         model.reset();
-        presenter.onCountButtonPressed(new CountView.CountButtonPressedEvent());
+        presenter.onCountButtonPressed();
         assertEquals(model.getCount(), 1);
         verify(view).setCount("1");
         verifyNoMoreInteractions(view);
@@ -47,11 +40,11 @@ public class PresenterTest {
 
     @Test
     public void isShouldResetCount() {
-        presenter.onCountButtonPressed(new CountView.CountButtonPressedEvent());
-        presenter.onCountButtonPressed(new CountView.CountButtonPressedEvent());
-        presenter.onCountButtonPressed(new CountView.CountButtonPressedEvent());
+        presenter.onCountButtonPressed();
+        presenter.onCountButtonPressed();
+        presenter.onCountButtonPressed();
         assertEquals(model.getCount(), 3);
-        presenter.onResetButtonPressed(new CountView.ResetButtonPressedEvent());
+        presenter.onResetButtonPressed();
         assertEquals(model.getCount(), 0);
         verify(view, times(4)).setCount(anyString());
         verifyNoMoreInteractions(view);
